STRY50557163 : [Delta-Loading]Find the delta of changes  between currently loaded application files and new changes applied while  Import from SC
[Delta-Loading]Find the delta of changes between currently loaded application files and new changes applied while Import from SC

This is a customization in the JGIT to filter out the noise and reduce the source control app install on the service now instance.

The customization ONLY includes changes to DiffCommand and DiffFormatter and we exposed a new method and the purpose is to not break existing JGIT in any way.